folder/backend/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
distlib==0.3.9
distro==1.9.0
Django==5.2
django-bootstrap5==25.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
executing==2.2.0
filelock==3.17.0
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jupyter_client==8.6.3
jupyter_core==5.7.2
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==7.0.0
psycopg2==2.9.10
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==308
pyzmq==26.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.29.2
wcwidth==0.2.13
Werkzeug==3.1.3
WTForms==3.2.1
